+++ compiler-core/src/error.rs
@@ -5287,7 +5287,10 @@ and cannot be accessed by other modules."
                 extra_labels: vec![],
             }),
         },
-        TypeError::ListPrependWithoutElements { location } => Diagnostic {
+        TypeError::ListPrependWithoutElements {
+            location,
+            tail_location: _,
+        } => Diagnostic {
             title: "Redundant list prepend".to_string(),
             text: "See: https://tour.gleam.run/basics/lists/".to_string(),
             hint: Some("`list` can be used directly instead of `[..list]`".into()),