+++ CHANGELOG.md
@@ -234,6 +234,24 @@
   renaming for function argument labels.
   ([Alistair Smith](https://github.com/alii))
 
+- The language server can now offer a "Remove redundant list prepend" code
+  action, for example:
+
+  ```gleam
+  const wibble = [1, 2]
+  const wobble = [..wibble]
+               // ^^^^^^^^ Trigger code action here
+  ```
+
+  becomes
+
+  ```gleam
+  const wibble = [1, 2]
+  const wobble = wibble
+  ```
+
+  ([0xda157](https://github.com/0xda157))
+
 ### Formatter
 
 ### Compiler Wasm API